# openapi.yaml
openapi: 3.0.3
info:
  title: LBC Rider Backend API
  description: |
    REST API for the LBC Express rider delivery platform and admin portal.

    ## Authentication

    | Audience | Login | Header |
    |----------|-------|--------|
    | **Rider** | `POST /api/auth/login` | `Authorization: Bearer <rider_jwt>` |
    | **Admin** | `POST /api/admin/auth/login` | `Authorization: Bearer <admin_jwt>` |

    ## Using Swagger UI

    1. Click the green **Authorize** button (top right).
    2. You will see two token fields:
       - **RiderAccessToken** — for rider APIs (manifest, delivery, route, etc.)
       - **AdminAccessToken** — for `/api/admin/*` APIs
    3. Paste the full header value: `Bearer <your_jwt_token>`
    4. Authorize only one at a time (both use the `Authorization` header).

    Public routes (no rider JWT): `/health`, `/api/auth/login`, `/api/admin/auth/*`, `/api/geocode/*`, `/api/docs`

    Admin routes additionally require an admin JWT via `adminMiddleware`.

    ## Base URL

    Default local port: `8000` (override with `PORT` env var).
  version: 1.0.0
  contact:
    name: LBC Rider Backend
servers:
  - url: http://localhost:8000
    description: Local development
  - url: http://localhost:3002
    description: Local (.env.example default)
...
paths:
...
  /api/route/optimize:
    post:
      tags: [Route]
      summary: Optimize stop sequence (Fleet Routing or local fallback)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [manifestId]
              properties:
                manifestId:
                  type: string
                priorityStopId:
                  type: string
                  example: stop-a1b2c3d4
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  newOrder:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
  /api/route/compute:
    post:
      tags: [Route]
      summary: Compute route between two points
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LatLngPair'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteComputeResult'
  /api/route/directions:
    post:
      tags: [Route]
      summary: Directions with optional waypoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [origin, destination]
              properties:
                origin:
                  $ref: '#/components/schemas/LatLng'
                destination:
                  $ref: '#/components/schemas/LatLng'
                waypoints:
                  type: array
                  items:
                    $ref: '#/components/schemas/LatLng'
      responses:
        '200':
          description: Google Routes API response
  /api/route/reorder:
    post:
      tags: [Route]
      summary: Manually reorder stops on manifest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [manifestId, stopOrder]
              properties:
                manifestId:
                  type: string
                stopOrder:
                  type: array
                  items:
                    type: string
                  description: Business stop IDs in desired order
      responses:
        '200':
          description: Order updated

  # ─── Geocode ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    RiderAccessToken:
      type: apiKey
      in: header
      name: Authorization
      description: |
        **Rider JWT** — enter the full value including the Bearer prefix.

        Example: `Bearer eyJhbGciOiJIUzI1NiIs...`

        Get your token from `POST /api/auth/login` → response field `token`.
    AdminAccessToken:
      type: apiKey
      in: header
      name: Authorization
      description: |
        **Admin JWT** — enter the full value including the Bearer prefix.

        Example: `Bearer eyJhbGciOiJIUzI1NiIs...`

        Get your token from `POST /api/admin/auth/login` → response field `token`.

        Note: Only authorize one token at a time (rider OR admin) since both use the same header.
...
  schemas:
...
    LatLng:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
        lng:
          type: number
    LatLngPair:
      type: object
      required: [origin, destination]
      properties:
        origin:
          $ref: '#/components/schemas/LatLng'
        destination:
          $ref: '#/components/schemas/LatLng'
...
    RouteComputeResult:
      type: object
      properties:
        encodedPolyline:
          type: string
        durationSeconds:
          type: integer
        distanceMeters:
          type: integer
        distanceKm:
          type: string
        eta:
          type: string
...
security:
  - RiderAccessToken: []